Moved Neural network examples to Neural_network class.
